+++ app/drivers/cd74hc406x/Src/cd74hc406x.c
@@ -17,7 +17,7 @@ const int select_pins[] = { CD74HCT4067_S0_Pin, CD74HCT4067_S1_Pin, CD74HCT4067_
  */
 static void cd74hc406x_select_channel(uint8_t channel){
 
-	for (uint8_t i = 0; i < 4; i++)
+	for (uint8_t i = 0; i < (uint8_t)4; i++)
 	{
 		if (channel & (1<<i))
 			HAL_GPIO_WritePin(GPIOB, select_pins[i], GPIO_PIN_SET);
@@ -28,12 +28,12 @@ static void cd74hc406x_select_channel(uint8_t channel){
 }
 
 
-void cd74hc406x_read_all_channels(uint32_t *channels){
+void cd74hc406x_read_all_channels(uint8_t *channels){
 
-	for(uint8_t i = 0; i < 16; i++){
+	for(uint8_t i = 0; i < (uint8_t) 16; i++){
 
 		cd74hc406x_select_channel(i);
-		channels[i] = HAL_GPIO_ReadPin(CD74HCT4067_S_GPIO_Port,CD74HCT4067_S_Pin);
+		channels[i] = (uint8_t) HAL_GPIO_ReadPin(CD74HCT4067_S_GPIO_Port,CD74HCT4067_S_Pin);
 
 	}
